+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 2024-06-16
+
+### Fixed
+
+- Removed erroneous world save for Shrewd Elemental Resist/Thick Skin.
+
 ## 2024-06-15
 
 ### Added